Rename server to api
